stac/requirements.txt
python-dotenv>=1.0.0
google-cloud-storage>=2.14.0
pypgstac>=0.7.0
pystac>=1.8.0
rio-stac>=0.5.0
rasterio>=1.3.0
boto3>=1.34.0
psycopg>=3.2.9
psycopg_pool>=3.2.6